Modify q_new(), q_free(), q_insert_*()
Add tail pointer and size variable for struct queue_t.
They will improve time complexity for q_insert_tail() and q_size().

q_free() deletes the list elements from head until the queue is empty.
Because that we should delete only at head in this queue.

q_insert_*() left an issue that strlen() may cause fault when there is
no '\0' in the copied string.
Assuming that '\0' exists, we copy one more char to include '\0'.
We do not have to append the char ourselves.