Modify q_sort() and add a string compare function
In a linked list, most elements cannot be accessed directly. And if I
construct an array version of the list, they will use too many memory
space. Therefore, I introduce an array of pointers to element, which
does not occupy that much space.
We can sort the array and relink the list according to the sorted array.
That completes the sort of our linked list.

The quick sort is cuurently using the right-most element as pivot. It
may be improved by choosing a better pivot.

The homework requirement did not mention how to compare. Thus I
chose one.

Beyond:
1) There seems to be a bug in q_insert_tail(), fix it.
2) Finish q_reverse().
3) Try to find a non-branch method for comparison.
4) Improve q_sort().