Rewrite q_sort()
malloc() is not used in q_sort() in this version and the code is more
simple and clear.
q_sort() is solved recursively by sortList(). In which, we cannot use
void-return-type companion with **head, otherwise we would dereference a
NULL pointer.
cmp is a parameter of the sortList() instead of directly using in
sortList() because a parameter is easier to change if we want use a
different ordering.

The pointer to pointer parameter in push() makes the code to be more
clear and more compatible with the ternary operator.